Refactored and combined into one file
